Rename client_key to identity_key
