+++ tools/nym-cli/src/validator/mixnet/operators/mod.rs
@@ -4,8 +4,8 @@
 use nym_cli_commands::context::ClientArgs;
 use nym_network_defaults::NymNetworkDetails;
 
-pub(crate) mod client_key;
 pub(crate) mod gateways;
+pub(crate) mod identity_key;
 pub(crate) mod mixnodes;
 pub(crate) mod name;
 pub(crate) mod services;
@@ -24,6 +24,6 @@ pub(crate) async fn execute(
         ) => mixnodes::execute(global_args, mixnode, network_details).await,
         nym_cli_commands::validator::mixnet::operators::MixnetOperatorsCommands::ServiceProvider(service) => services::execute(global_args, service, network_details).await,
         nym_cli_commands::validator::mixnet::operators::MixnetOperatorsCommands::Name(name) => name::execute(global_args, name, network_details).await,
-        nym_cli_commands::validator::mixnet::operators::MixnetOperatorsCommands::IdentityKey(identity_key) => client_key::execute(global_args, identity_key, network_details).await,
+        nym_cli_commands::validator::mixnet::operators::MixnetOperatorsCommands::IdentityKey(identity_key) => identity_key::execute(global_args, identity_key, network_details).await,
     }
 }
